requirements.txt
black==20.8b1
coverage==5.5
Django==3.1.7
djangorestframework==3.12.2
factory-boy==3.2.0
Faker==6.5.0
flake8==3.8.4
ipdb==0.13.5
isort==5.7.0
pre-commit==2.11.0
psycopg2-binary==2.8.6
pyflakes==2.2.0
python-decouple==3.4
requests==2.25.1
sendgrid==6.6.0
